ml/requirements.txt
xgboost==2.0.3
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
seaborn==0.13.2
shap==0.45.0
joblib==1.4.2
requests==2.31.0
python-dotenv==1.0.1
